sleepmapper/requirements.txt
torch
torchaudio
librosa
numpy
pandas
scikit-learn
matplotlib
tqdm
onnx
onnxruntime
wfdb
transformers
pyyaml